Add the first LAVA v2 templates

+++ lava-kernel-ci-job-creator.py
@@ -89,7 +89,7 @@ def create_jobs(base_url, kernel, plans, platform_list, targets, priority):
                         job_name = tree + '-' + kernel_version + '-' + defconfig[:100] + '-' + platform_name + '-' + device_type + '-' + plan
                         job_json = directory + '/' + job_name + '.json'
                         template_file = cwd + '/templates/' + plan + '/' + str(template)
-                        if os.path.exists(template_file):
+                        if os.path.exists(template_file) and template_file.endswith('.json'):
                             with open(job_json, 'wt') as fout:
                                 with open(template_file, "rt") as fin:
                                     for line in fin:
